third_party/WebKit/LayoutTests/http/tests/appcache/404-manifest-expected.txt

Test that subresources can be loaded if manifest is not available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gotCheckingEvent is true
PASS window.applicationCache.status is applicationCache.UNCACHED
PASS event.reason is "manifest"
PASS event.url is "http://127.0.0.1:8000/appcache/does-not-exist.manifest"
PASS event.status is 404
PASS 'message' in event is true
PASS req.responseText is "Hello, World!"
PASS successfullyParsed is true

TEST COMPLETE
